Oss release 26.0.4 created 2020-09-30-19-57
see CHANGELOG.txt for details

Original commit sha: 9c00232db90d04b9a9453507e4699ab02a2cbef8

+++ client/ramses-client/ramses-text-api/UtfUtils.cpp
@@ -127,7 +127,7 @@ namespace ramses
             }
             else
             {
-                LOG_ERROR(CONTEXT_TEXT, "UtfUtils::ConvertCharUtf32ToUtf8: invalid Unicode " << convertChar << " Skipping Character\n" );
+                LOG_ERROR(CONTEXT_TEXT, "UtfUtils::ConvertCharUtf32ToUtf8: invalid Unicode " << static_cast<uint32_t>(convertChar) << " Skipping Character\n" );
             }
 
             return utf8string;
